+++ forum-common/src/main/java/org/jh/forum/common/dto/response/ListAnnouncementResponse.java
@@ -5,7 +5,7 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 
 /**
- * 查询公告列表响应DTO
+ * 查询公告列表响应DTO（管理员版本）
  * 
  * @author SituChengiang
  */
@@ -17,9 +17,8 @@ public class ListAnnouncementResponse extends BaseListResponse<ListAnnouncementR
     @Data
     @Schema(description = "公告列表项")
     public static class AnnouncementItemResponse {
-
         @Schema(description = "公告ID", example = "1")
-        private Integer id;
+        private Long id;
 
         @Schema(description = "公告标题", example = "重要系统维护通知")
         private String title;
